Implementing Basic Level-Editor.

+++ IH11-Game/src/level/graphics/LevelTexture.java
@@ -15,16 +15,16 @@ public class LevelTexture {
 	private BufferedImage[] tileArray;
 	private BufferedImage image;
 
-	public static final int NORTHWEST = 0;
-	public static final int NORTH = 1;
-	public static final int NORTHEAST = 2;
-	public static final int WEST = 3;
-	public static final int CENTER = 4;
-	public static final int EAST = 5;
-	public static final int SOUTHWEST = 6;
-	public static final int SOUTH = 7;
-	public static final int SOUTHEAST = 8;
-	public static final int AIR = 9;
+	public static final byte NORTHWEST = 0;
+	public static final byte NORTH = 1;
+	public static final byte NORTHEAST = 2;
+	public static final byte WEST = 3;
+	public static final byte CENTER = 4;
+	public static final byte EAST = 5;
+	public static final byte SOUTHWEST = 6;
+	public static final byte SOUTH = 7;
+	public static final byte SOUTHEAST = 8;
+	public static final byte AIR = 9;
 
 	public LevelTexture() {
